requirements.txt
flask==2.0.1
Jinja2==3.0.1
reportlab==4.0.4
Pillow==10.0.0
langchain
langchain-groq==0.1.0
python-dotenv==1.0.0
pymongo==4.6.1
bcrypt==4.1.2
PyJWT==2.8.0
flask-cors==4.0.0